extras/file-converter/requirements.txt
# File Converter Dependencies

# Core dependencies
Pillow>=10.0.0
pytesseract>=0.3.10
python-docx>=1.0.0
pdfplumber>=0.10.0
openpyxl>=3.1.0
html2text>=2020.1.16
chardet>=5.0.0

# UI
tkinterdnd2>=0.3.0

# Build (optional - only needed for creating .exe)
# pyinstaller>=6.0.0
